+++ src/common/gitProvider/azureDevops.ts
@@ -2,4 +2,8 @@ import { GitProvider } from ".";
 
 export class AzureDevopsProvider extends GitProvider {
 
+    getLabel(): string {
+        return "sfdx-hardis Gitlab connector";
+      }
+
 }