+++ src/common/gitProvider/gitProvider.d.ts
@@ -1,0 +1,10 @@
+export declare type PullRequestMessageRequest = {
+    title: string;
+    message: string;
+    status: ["valid", "invalid", "tovalidate"]
+}
+
+export declare type PullRequestMessageResult = {
+    posted: boolean;
+    providerResult: any
+}