+++ src/common/gitProvider/github.ts
@@ -2,4 +2,8 @@ import { GitProvider } from ".";
 
 export class GithubProvider extends GitProvider {
 
+    getLabel(): string {
+        return "sfdx-hardis Gitlab connector";
+      }
+
 }