+++ src/common/gitProvider/index.ts
@@ -1,10 +1,43 @@
+import { SfdxError } from "@salesforce/core";
+import * as c from "chalk";
+import { isCI, uxLog } from "../utils";
+import { AzureDevopsProvider } from "./azureDevops";
+import { GithubProvider } from "./github";
+import { GitlabProvider } from "./gitlab";
+import { PullRequestMessageRequest, PullRequestMessageResult } from "./gitProvider";
+
 export abstract class GitProvider {
-  protected host: string;
+
+  protected serverUrl: string;
   protected token: string;
 
-  constructor(host: string, token: string) {
-    this.host = host;
-    this.token = token;
+  static getInstance(): GitProvider {
+    // Azure
+    if (process.env.SYSTEM_ACCESSTOKEN) {
+      return new AzureDevopsProvider();
+    }
+    // GitHub
+    else if (process.env.CI_JOB_TOKEN) {
+      return new GithubProvider();
+    }
+    // Gitlab
+    else if (process.env.GITHUB_TOKEN) {
+      return new GitlabProvider();
+    }
+    else if (isCI) {
+      uxLog(this, c.grey("To use sfdx-hardis GitProvider capabilities, SYSTEM_ACCESSTOKEN, CI_JOB_TOKEN or GITHUB_TOKEN must be accessible for Azure Pipelines, Gitlab or GitHub"))
+    }
+    return null;
+  }
+
+  getLabel(): string {
+    throw new SfdxError("getLabel should be implemented on this call");
+  }
+
+  async postPullRequestMessage(prMessage: PullRequestMessageRequest): Promise<PullRequestMessageResult> {
+    uxLog(this, c.yellow("Method postPullRequestMessage is not yet implemented on " + this.getLabel() +
+      " to post " + JSON.stringify(prMessage)));
+    return null;
   }
 
   async getParentMergeRequestId(): Promise<number> {
@@ -12,14 +45,17 @@ export abstract class GitProvider {
   }
 
   async getPipelineId(): Promise<string> {
-    return "This method is not yet implemented.";
+    uxLog(this, c.yellow("Method getPipelineId is not yet implemented on " + this.getLabel()));
+    return null;
   }
 
   async getJobsFromPipeline(): Promise<string> {
-    return "This method is not yet implemented.";
+    uxLog(this, c.yellow("Method getJobsFromPipeline is not yet implemented on " + this.getLabel()));
+    return null;
   }
 
   async getDeployId(): Promise<string> {
-    return "This method is not yet implemented.";
+    uxLog(this, c.yellow("Method getDeployId is not yet implemented on " + this.getLabel()));
+    return null;
   }
 }